# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Techhype-API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "JWT": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
...
    "/address/{profileId}": {
      "get": {
        "tags": [
          "Address"
        ],
        "description": "get address for a card profile.",
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Address"
        ],
        "description": "Add or update an address for a card profile.",
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Address"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "Address"
        ],
        "description": "Delete an address for a card profile.",
        "parameters": [
          {
            "name": "profileId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/address/{addressId}": {
      "get": {
        "tags": [
          "Address"
        ],
        "description": "get address By using the address ID.",
        "parameters": [
          {
            "name": "addressId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Address": {
      "type": "object",
      "properties": {
        "country": {
          "type": "string",
          "example": "Norway"
        },
        "street": {
          "type": "string",
          "example": "Vestre Holbergsallmenningen 15"
        },
        "postalCode": {
          "type": "string",
          "example": "3002"
        },
        "state": {
          "type": "string",
          "example": "Vestland"
        },
        "city": {
          "type": "string",
          "example": "Bergen"
        }
      },
      "required": [
        "country",
        "street",
        "postalCode",
        "state",
        "city"
      ]
    },
...
  },
  "security": [
    {
      "JWT": []
    }
  ]
}